# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ForbiddenError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BacklogCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"branchId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/CategoryCompleteOut"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"branchId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BranchNotFound": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"SessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session-id"
			},
			"JWTAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access-token"
			}
		}
	},
	"info": {
		"title": "Menuchi-API",
		"version": "1.0.0",
		"description": "Menuchi API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Mohammad Hosseini"
		}
	},
	"paths": {
...
		"/menus/backlog/{backlogId}": {
			"get": {
				"operationId": "GetBacklog",
				"responses": {
					"200": {
						"description": "Backlog is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4049 BranchNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BranchNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a backlog by ID.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/DefaultString"
						}
					}
				]
			}
		},
		"/menus": {
			"post": {
				"operationId": "CreateMenu",
				"responses": {
					"201": {
						"description": "Menu created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMenuCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4049 BranchNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BranchNotFound"
								}
							}
						}
					}
				},
				"description": "Creates a new menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMenuCompactIn"
							}
						}
					}
				}
			}
		},
		"/menus/branch/{branchId}": {
			"get": {
				"operationId": "GetAllMenus",
				"responses": {
					"200": {
						"description": "Menus are retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MenuCompleteWithCountsOut"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Retrieves all branch menus.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "branchId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/menus/{menuId}": {
			"get": {
				"operationId": "GetMenu",
				"responses": {
					"200": {
						"description": "Menu is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuCompletePlusOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4048 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a complete menu by ID.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMenu",
				"responses": {
					"204": {
						"description": "Menu updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4048 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					},
					"422": {
						"description": "4228 MenuValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuValidationError"
								}
							}
						}
					}
				},
				"description": "Updates a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuCompactIn"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenu",
				"responses": {
					"204": {
						"description": "Menu deleted successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/menus/{menuId}/cylinders": {
			"post": {
				"operationId": "CreateCylinder",
				"responses": {
					"201": {
						"description": "Cylinder created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateCylinderCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4048 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError -> A cylinder with the same provided days are already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4226 CylinderValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CylinderValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new cylinder in a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CylinderCompactIn"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "ReorderCylinders",
				"responses": {
					"204": {
						"description": "Cylinder orders in the menu updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "All menu cylinder IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders cylinders.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}/categories": {
			"post": {
				"operationId": "CreateMenuCategory",
				"responses": {
					"201": {
						"description": "Menu Category created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMenuCategoryCompleteOut"
								}
							}
						}
					},
					"400": {
						"description": "All item IDs must belong to the specified category.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4048 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError -> A menu category with the same name already exists within the specified cylinder.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4227 MenuCategoryValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuCategoryValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new menu category.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuCategoryCompactIn"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenuCategory",
				"responses": {
					"204": {
						"description": "Menu Categories deleted successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes menu categories.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "ReorderMenuCategories",
				"responses": {
					"204": {
						"description": "Menu category orders in the cylinder updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "All menu category IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders menu categories.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/menus/{menuId}/items/{menuItemId}/hide/{isHide}": {
			"patch": {
				"operationId": "HideMenuItem",
				"responses": {
					"204": {
						"description": "Menu Item hide/unhide successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Hide/Unhide a menu item.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "menuItemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "isHide",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}